requirements.txt
aiohttp==3.7.2
arrow==0.17.0
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.11.8
chardet==3.0.4
idna==2.10
multidict==5.0.2
python-dateutil==2.8.1
requests==2.25.0
six==1.15.0
slackclient==2.9.3
typing-extensions==3.7.4.3
urllib3==1.26.2
yarl==1.6.3
